both obj store and synthetic tests working
