+++ tests/gateway/test_gateway_synthetic.py
@@ -4,84 +4,19 @@ from skyplane.api.config import TransferConfig
 import time
 import docker
 import argparse
-from skyplane.obj_store.object_store_interface import ObjectStoreInterface, ObjectStoreObject
 from skyplane.api.config import TransferConfig
 from skyplane.utils.fn import PathLike
 from skyplane.api.dataplane import Dataplane
 import os
-from skyplane.compute.server import Server
+from skyplane.cli.impl.progress_bar import ProgressBarTransferHook
 from skyplane.planner.planner import MulticastDirectPlanner
 from skyplane.api.transfer_job import CopyJob, TestCopyJob
 from skyplane.api.provisioner import Provisioner
 import subprocess
 
+from tests.gateway.test_server import TestServer
+from tests.gateway.test_interface import TestInterface
 
-class TestObject(ObjectStoreObject):
-    def full_path(self):
-        return f"test://{self.bucket}/{self.key}"
-
-class TestInterface(ObjectStoreInterface):
-    """Test interface simulates an object store interface"""
-
-    def __init__(self, region_tag, bucket_name):
-        self.bucket_name = bucket_name
-        self._region_tag = region_tag
-        self.provider = "test"
-
-    def path(self):
-        return f"{self.provider}://{self.bucket_name}"
-
-    def bucket(self):
-        return self.bucket_name
-    
-    def bucket_exists(self) -> bool:
-        return True
-
-    def region_tag(self):
-        return self._region_tag
-    
-    def list_objects(self, prefix="") -> Iterator[Any]:
-        for key in ["obj1", "obj2", "obj3"]: 
-            obj = self.create_object_repr(key)
-            obj.size = 100 
-            yield obj
-
-    def get_obj_size(self, obj_name) -> int:
-        return 100
-
-    def get_obj_last_modified(self, obj_name):
-        return "2020-01-01"
-
-    def get_obj_mime_type(self, obj_name):
-        return None
-
-    def download_object(
-        self, src_object_name, dst_file_path, offset_bytes=None, size_bytes=None, write_at_offset=False, generate_md5: bool = False
-    ) -> Tuple[Optional[str], Optional[bytes]]:
-        return
-
-    def upload_object(
-        self,
-        src_file_path,
-        dst_object_name,
-        part_number=None,
-        upload_id=None,
-        check_md5: Optional[bytes] = None,
-        mime_type: Optional[str] = None,
-    ):
-        return
-
-    def delete_objects(self, keys: List[str]):
-        return
-
-    def initiate_multipart_upload(self, dst_object_name: str, mime_type: Optional[str] = None) -> str:
-        return
-
-    def complete_multipart_upload(self, dst_object_name: str, upload_id: str) -> None:
-        return
-    
-    def create_object_repr(self, key: str) -> TestObject:
-        return TestObject(provider="test", bucket=self.bucket_name, key=key)
 
 """
 
@@ -90,74 +25,6 @@ This test module creates local docker containers that run gateway code.
 For testing, we create an artificial region call "test", and create a TestInterface (the ObjectStore interface) that simulates interacting with an object store, and 
 a "test:region" region tag that identifies artificial regions to deploy gateway docker containers to. 
 """
-
-
-class TestServer(Server):
-
-    """Test Server runs a gateway container locally to simulate deployed gateways"""
-
-    def __init__(
-        self, region_tag, log_dir=None, auto_shutdown_timeout_minutes: Optional[int] = None, local_port=None, gateway_api_url=None
-    ):
-        print("CREATE TEST SERVER")
-        super().__init__(region_tag, log_dir, auto_shutdown_timeout_minutes)
-        self.command_log = []
-        self.gateway_log_viewer_url = None
-        self.gateway_api_url = None
-        self.init_log_files(log_dir)
-        self.ssh_tunnels: Dict = {}
-        self.local_port = local_port
-        self.gateway_api_url = f"http://127.0.0.1:{self.local_port}"
-
-    def run_command(self, command) -> Tuple[str, str]:
-        # execute command locally
-        print("command:", command)
-        result = subprocess.run(command, shell=True, capture_output=True, encoding="utf8")
-        print("resutl", result, result.stderr, result.stdout)
-        return result.stdout, result.stderr
-
-    def download_file(self, remote_path, local_path):
-        """Pretend to download a file from the server"""
-        # Create the necessary directories in the local path
-        local_directory = os.path.dirname(local_path)
-        os.makedirs(local_directory, exist_ok=True)
-
-        # Copy the file from the remote path to the local path
-        shutil.copy2(remote_path, local_path)
-
-    def upload_file(self, local_path, remote_path):
-        """Pretend to upload a file to the server"""
-        # shutil.copyfile(local_path, remote_path)
-
-        # Create the necessary directories in the local path
-        remote_directory = os.path.dirname(remote_path)
-        if remote_directory != remote_path:
-            os.makedirs(remote_directory, exist_ok=True)
-
-        # Copy the file from the remote path to the local path
-        shutil.copy2(local_path, remote_path)
-
-    def write_file(self, content_bytes, remote_path):
-        """Write a file on the server"""
-        print(remote_path)
-        with open(remote_path, "wb") as f:
-            f.write(content_bytes)
-
-    def copy_public_key(self, pub_key_path: PathLike):
-        """Pretend to append public key to authorized_keys file on server."""
-        pass
-
-    def uuid(self):
-        return f"test:{self.region_tag}"
-
-    def public_ip(self):
-        return f"localhost:{self.local_port}"
-
-    def private_ip(self):
-        return f"localhost:{self.local_port}"
-    
-    def instance_name(self):
-        return self.region_tag
 
 
 def wait_container_running(container_name):
@@ -203,7 +70,6 @@ def run(gateway_docker_image: str, restart_gateways: bool):
     job._dst_ifaces = [TestInterface("test:dest_region", "dest")]
 
     topology = MulticastDirectPlanner(1, 64, TransferConfig()).plan([job])
-    print([g.region_tag for g in topology.get_gateways()])
     # print(topology.generate_gateway_program("test:source"))
 
     source_server = TestServer(topology.src_region_tag, local_port=8081)
@@ -251,7 +117,7 @@ def run(gateway_docker_image: str, restart_gateways: bool):
 
     # run dataplane
     try:
-        dataplane.run([job])
+        dataplane.run([job], hooks=ProgressBarTransferHook(topology.dest_region_tags))
     except KeyboardInterrupt:
         print("Keyboard interrupt: copying logs...")
     except Exception as e:
